requirements.txt
Django
Pillow
django-jazzmin
django-extensions
django-tinymce